Yes, I did change this checkin, although it was totally correct. So why, you may ask???
Baaasically (Hi Alex) the same change that arigo did,
but I really had to add some commentary,
which should be dropped, after we fulfilled
our task of really documenting and supporting
the maintainers of tcc, who were so very helpful
for us. Please, try to digest that sentence.

We need to keep track of actions which we
want to propagate through the pypy environment.
We are an open-sourced, EU-funded project.
In this outstanding situation, we are oblieged
not to just digest and valuate existing other OS
projects, but we have to support them, I guess.

Now, by putting tcc into a real-world situation,
I think that we also added to its reasons why it
should exist, of course.
But if we do use it, and if we do find problems
with it, then it should be a consequence to support
it and to provide follow-ups and bug-reports.

Was this spelled correctly, guys?

cheers - chris

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4010655