a CPython like __repr__ for methods (new-style).
it doesn't return '<built-in...' methods for
e.g. repr({}.items) though because we don't
really have the distinction between builtin-in
and not builtin (except for the hack when
we export builtin functions from the builtins
module).  So we either need to extend that
hack or to change the test_repr.py i guess ...

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4012066