issue39 resolved
Brain-stormed short mission statement for the front page.  We tried to make it
very self-contained, and very short.  Now a link, "more...", points to an
extended version which gives the main goals.  The links to the existing
projects -- mainly Psyco and Stackless -- are moved in the next paragraph,
as well as to the Related projects section.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4012642