More strangeness, more progress: test_rbuiltin.test_instantiate_multiple().
instantiate() is now a "built-in" from the point of view of the annotator,
which is progress.

All vtables have an entry 'instantiate' that points to a function that
instantiates the corresponding class.  Building this function is actually "a
bit" messy.  Then instantiate() can just be implemented by calling it.

We are reusing a trick we already used in the previous check-in: here, the
function that instantiate a class are defined (in normalizecalls) as

    def my_instantiate():
        return instantiate(cls)

which means that instantiate() is actually implemented as follows: if the
argument is a known constant class, we instantiate it in the normal way
(rclass.py, rtype_new_instance()).  Otherwise, we load and call the
appropriate version of my_instantiate(), which itself contains again a call to
instantiate() -- but this time with a constant 'cls'...

For reference, the previous check-in used the same trick for calling classes.
A call to a constant class was already implemented some time ago, so a call to
a non-constant class can be done by calling a helper function __new__() that
is defined as calling a constant class.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4014566