enabling marshal:
had to initialize buffer in the exception case of mul_string_times.
This was strange, because this was not consequently always needed.
There seems to be some non-determinism.

I found no way to reliably reproduce this. When finally trying
to remove it, the problem came up, again, so I kept it.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4015717