a little clean-up to int/long/floatobject:
Use multimethod exceptions instead of directly
calling surrogate functions.
Reason:
I would like to have the objspace implementations
create exactly the type they are made for, or raise
NonImplemented. That makes it easier to call such
functions from somewhere else. Although I also think
that calling into objspace implementations is just
a temporary thing, and we should have an extra sub-interface
for everything that gets called "under the hood" and
refrain from calling into W_Anything if it is avoidable.

I will employ this stylepretty soon, isolating the long impl
from the objspace and putting it into,say, longutil for the moment.

Btw., the way pow() needs to be redirected is slightly inefficient,
and using the NotImplementedto resolve it is slightly an abuse,
because the reason why we cannot early dispatch on it is that we are
lacking an ability to express an interface with non-negative ints.
Would it be possible to do that? Like pow__Int_Unsigned_None ???

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4016016