Support for "external types" for annotation and rtyping, following the same lines as external functions.  Some types like thread.LockType can be declared in the new "typetable" of rpython.extfunctable.  These types must have a simple interface -- basically, only methods, which are themselves external functions.
The annotator uses a new SomeExternalObject to track objects of one of these
external types.  The rtyper turns them into pointers to opaque data.  There is
some changes all over the place to support this kind of usage of opaque data;
I'm not sure yet if it's really nice to use opaque types for this, because I'm
unsure how to support the result in genc.  (The problem is that this gives
pointers to opaque objects that still need the same memory management as
pointers to regular structs...)

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4016172