Another "bug" that showed up after making low level list length explicit. We now need to pass all the lengths explicitly.
I think we are lacking some more abstract interface, here.
Instead of referring to properties of other primitives,
directly, we should have a set of standard accessors which do the
right thing, regardless whether we change layout or not.

For instance, lltype.length always means the current length of
an object, whether this is the length of an underlying structure,
or the length the objects wants to pretend.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4017160