About this pseudo-memory-leak -----------------------------
dropping frame.last_exception after leaving it.

This solves the problem that we get memory leaks on every
applevel expcetion if we have to GC at hand. This is the
only obvious circle that we have, and it is also not necessary
and not conforming to the CPython implementation,
which too clears exception state from frames when they are
not executing.

One might argue that I should have put this into pyframe.py
instead of eval; feel free to do so.
My reason was: the eval loop of pyframe is already very deeply
nested, and I would have to add another finally clause, which not
onlymade it less readable, but gave a measurable performance hit.
Putting it into the resume method is cheap, because we have
a finally clause there, anyway.

It does logically not really belong there, so I could understand
if somebody's aesthetics is disturbed and he moves it to frame.eval

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4017168