this check-in is just to show a funny observation.
When a variable is conditionally assigned two different ranges,
the ranges are kept, unless the steps are different.
Well, after all this is not surprizing.
But I thing the action taken is not pleasant:
the range is turned into a list in this case!

I guess what we want is either special-casing the respective blocks,
or at least map this to the variable step case, if this is possible.

Please let me know what you think.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4017487