executable                  abs.richards   abs.pystone   rel.rich   rel.pystone pypy-c-17439                  35165 ms         668.586      42.4           61.1 pypy-c-17600                  26388 ms         900.481      31.8           45.4 pypy-c-17634                  20108 ms        1017.720      24.2           40.1 pypy-c-17649                  22662 ms        1035.910      27.3           39.4 pypy-c-17674-nolowmem         15817 ms        1981.470      19.1           20.6 pypy-c-17674-t-lowmem         16834 ms        1274.650      20.3           32.1 python 2.3.3                    830 ms       40861.400       1.0            1.0
17649 was with explicit fixed stack.
Changes after 17634 we not included.
17674 has an outrageous effect. I cannot really
find out what it was. Did Armin do the fixed stack
patch already? Probably not. Was it Samuele's avoiding
of duplicate zeroing? Really just that? I think so, and
this is incredible.
Even more incredible is the fact that not using using
t-lowmem accelerates pystone so much. This is an indicator
that we missed something used in pystone that still contains
applevel code. I can't believe it, will find it tomorrow.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4017675