added some provision to get the complete callgraph.
The problem was that we do some initial callgraph,
but in the later transformations, we don't update that,
and it would be not so trivial to enforce this.

Therefore, I created a property complete_callgraph,
which is recomputed whenever new graphs are added.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4019298