trying to see the difference in calling just __new__.
We forgot about this but there's unpacking and repacking of Arguments with __args__ signatures,
going on.

Also the logic now in CALL_FUNCTION is suboptimal because in the cases with 0-4 arguments it will
fallback to Arguments, instead of ArgumentsFromValuestack, so the latter is a bit wasted.

Our gateway should become more clever about the __args__ case if we want more speed.
Arguments will not help with that because breaking out the first argument is involved.

Also Arguments is a nice abstraction in theory but is a bit too heavyweight, and the last refactoring
didn't improve that at all with three/two versions of the same subtle code.

Arguments is nice for the analysis and for the most complex cases (*, ** etc) and error handling...

it would be best to have just the most complex version around and use lighter approaches for simpler
cases, this means rethinking the call multimehtod and its __call__ Function/gateway, the latter
is involved in the __args__ signature issue too.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4022965