(cfbolz, pedronis):
added an implementation of uthreads that use stackless and interpreter-level
coroutines. seems to be translatable and it is possible to run the examples.
Sidenode: it seems to be possible to use greenlets to implement
interpreter-level coroutines at least to some extend.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4024216