Add a new lltype function: cast_subarray_pointer().  Allows casting from a regular Array to a FixedSizeArray that represents a slice of the original Array.  (Motivation: rctypes.)
Implementation-wise, this requires a new class _subarray in lltype.py.
There is no new ll operation to support this, as it is rare and we
shouldn't burden the back-end about it.  Instead, it can be done with
a sequence of address manipulations.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4025727