fix propagation of exception out r_dict operations.
Notice that this required having the flow space assume that getitem/setitem/delitem can raise any exceptions.
Changes with some tests at various levels to cope with this.

Notice that the pruning of excepts links when a getitem appears in an unrelated try/except is less strong as a
consequence.

Now {}[[]] fails with a TypeError as expected in a translated PyPy.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4025965