Half refactoring of the AttrFamily: it is unchanged for FrozenDescs, but for ClassDescs it now records attribute access for each attribute name independently.  See new test in test_rpbc.
This is prompted by PyPy translation crashes: introducing a memo
function on the pbc of all W_Root subclasses forced them to be all in
the same AttrFamily, which had terribly bad effects, as follows: a few
subclasses of ast.Node are instantiated generically (by calling a
variable that can be any of them).  The same is true for the subclasses
of PyFrame.  However, the __init__ in these two parts of the hierarchy
are completely different.  So any common pbc attribute on all W_Root
subclasses caused the __init__ to come together too => obsure crash.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4026822