ah ah. too funny.
started working on making the semispace gc tested with the transformer. tried to use test_gc to see
if I broke something. there were some perplexing problems, discovered that the maze of setup and teardown
in test_gc made it so that mark and sweep was not tested by it at all anymore and it shared some
of the tweaked semispace problems. ah ah, that's why no prints.
of course mark and sweep was broken in test_gc: gcwrapper trying to annotate the __init__ which is not rpython
anymore, to shallow typos, to the os.write and strings requiring in theory that the gc function graphs
be converted too.

after quite a bit of pushing and pulling test_gc is testing successfully mark and sweep again.
Of course accumulated more reasons for wanting to kill it and gcwrapper.

ah ah. too funny. not much progress on the semispace transformer testing tough :(

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4027380