(antocuni, arigo)
Handle List of Void correctly.

The problem is that .NET don't allow us to use the void type as a real
type, so we have to workaround this. The solution for List(Void) is to
have a special-cased class whose methods don't take care of parameters
of ITEMTYPE type, then we special-cased some place for taking care of
this.

Probably in the future we will need a similar solution for Dict with
Void keys or values, and so on.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4028455