(antocuni, arigo)
override:ignore of methods is too difficult to implement in
ootypesystem and not so useful because we can use we_are_translated
instead, so we say it's not supported.

Remove the corresponding test and fix the only place it was used in
the interpreter.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4029686