This should fix the support for binary/text files both on unix and windows. There were many problems:
  - the .NET framework seems not to have any support for CRLF
    terminated text files, so we need to handle them explicitly;

  - we can't use StreamReader and StreamWriter because the are
    designed to handle unicode strings: the default encoding is UTF8,
    which breaks things when writing chars > 0x7F. The ASCII encoding
    doesn't handle chars > 0x7F too, so we can't use it.

The solution is to directly read bytes and casting them to
chars. Moreover, we need to pay attention on Windows when opening
files in text mode, because we need the CRLF conversion. We will
probably need a similar thing for MAC < OS9 in the future.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4030945