(pedronis, arigo)
Add a form of propagation of green variables, for the cases that don't
make merge points fail to unify in the timeshifted graphs because of it.

XXX needs tons of comments in the source code to explain *how* it does
this.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4032119