(pedronis, arigo)
Ouah!  A whole evening of installing Electric Fence and then
Valgrind to get at the roots of this problem.  Trivial, as
usual: the unsafe_call operation is not considered as a call,
so the GC roots are not pushed/popped around it.  Bouh.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4034750