Relatively large refactoring to implement support for choosing a revision (or a range of revisions) to build. The client still doesn't actually update to the revision, so it's not entirely functional yet, but it's all there.
I changed the way the data is passed around: instead of passing a tuple with
info from the startcompile script to the server, which passes it to the client,
a special object (BuildRequest) is passed around. This object contains info
about the email address, the revision and revision range, etc. When done
compiling, this object is serialized and stored on the BuildPath along with
the rest of the data.

The RequestStorage is removed, since the caching it did is no longer possible
(revision range spoils this).

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4035679