Better precision for when to generate bound checking versions of getitem. Have a transformation that rename getitem into getitem(_idx)(_key) based on whether the getitem is effectively guarded by such exception (IndexError, KeyError) exits.
Change annotation and rtyping to work accordingly.

Done with more precision for lists and strings for now, fallbacks in place for the rest.
(Should be done for setitem too)

Tests for the transformation in objspace/flow and for the final behavior in the rpython/test.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4036008