Fix the signed division and modulo code in the i386 backend.  At least, I assume that they are fixed, because the tests pass, but the code is hairy :-)
Added tests for the case of a constant divisor, which I had to do
differently because of lack of registers.  It makes sense anyway to test
this specifically, as they are likely to be special-cased in a more
optimizing version of the backend (e.g. replacing "/ 8" with ">> 3").

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4036825