Add the code needed for java code to throw exceptions.
We generate an instance of the Interlink interface which
has the various methods required, and assign it to a static
variable in the PyPy class.  This has the downside of only
allowing one concurrent PyPy interpreter: we could use thread-local
data instead (or pass the object as a parameter, but that would be
more of a pain).

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4037135