Add flush-on-exit capability to all our file objects.
Done with a new internal dict sys.pypy__exithandlers__.  In
space.finish(), we pop key/value pairs from this dictionary and just
call the value, as long as the dict is not empty.  This is probably
generally useful for situations where a number of small things must be
"finished" in any order, but finishing them might possibly create more
stuff that must itself be "finished" too.  The fact that it's a
dictionary means that it's cheap to add and remove many things from it.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4038737