Experimental check-in which could be reverted.  This makes the method cache global instead of per-thread.
It makes a pypy-c-thread almost as fast as a pypy-c-no-thread, when
threads are not used (<2% difference).  The reason is that
getexecutioncontext() needs to call RPyThreadIdent(), which appears to
be the main cause of a global slow-down of 4-5%.  I don't know if there
are real-world cases where the method cache would get significantly
trashed by being shared between threads.

An alternative to this check-in would be to improve
getexecutioncontext(), e.g. by replacing the dict {thread_id:
executioncontext} with a single executioncontext reference which is
carefully saved and restored when the GIL is released.  For now it would
be too much cleverness for too little gain, so I'll skip.

--HG--
branch : dist
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4055652