issue371 resolved
Merge the async-del branch:

  * Don't call the app-level __del__() methods when the gc calls the
    interp-level __del__() methods; instead, queue them for later.
    This fixes potential segfaults (see test_listobject.py).

  * Refactor the Action system.  The goal is to let actions like
    calling __del__() or delivering a signal occur immediately, i.e.
    before the next opcode, instead of having to wait for
    sys.checkinterval opcodes.

  * Use this to detect signals immediately; removes the need for
    a hack in module/_file's wrap_oserror_as_ioerror().

  * Whack at threads (most notably, fix a bogus test).

  * A few new tests relating to the above features.

The new Action classes are a bit more static (and quite more efficient,
reducing overhead).  The idea is to register during space initialization
which async actions can possibly occur, and allocate one bit of an
integer to each of them; in this way, the ExecutionContext's
bytecode_trace() can check if there is anything special to do or not
just by checking if this integer is != 0.

Line-by-line tracing is now done as an action too, which also reduces
the overhead when it is not used.

--HG--
branch : dist
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4056154