(iko,pedronis) Some test food for static data info. These generate insane results:
for targettestdicts.py, only the size of the dicts are counted, not
the contained strings (so the result is too small).

for targettestlistvsdict.py, the list only seems to count the size
of the contained strings, not the size of the array. The dict is
counted (correctly? since the same strings are in the list) as the
size of its structure.

--HG--
branch : dist
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/dist%4059062