another big refactoring: put part of InstanceNode responsibility into the new InstanceValue class.
InstanceNode and InstanceValue are similar: the main difference is that the
former is used exclusively by the LoopSpecializer, while the latter is used
exclusively by the LoopOptimizer.

This way, we make sure that their uses are not confounded by mistake, as I
think it happens in the old optimize.py (e.g. by using the cls attribute both
in the specialization phase and in the optimizing phase, but for slightly
different purposed)

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/branch/pyjitpl5-experiments%4065606