Merge the gc-compress branch.  This adds a new option, "--gcremovetypeptr", which compiles a framework GC in such a way as needing only one word to store the GC headers (including the GC-specific type data pointer) and the 'typeptr' field of all regular RPython instances.
The option is False by default for now, but with a bit more testing it
should become True.  It seems to give no noticeable speed-up, but no
slow-down either, and it saves memory.

Implemented with a new concept in rpython/lltypesystem, a "group", which
regroups otherwise-independent prebuilt structures in a lazy manner, and
offers unsigned 16-bits symbolics representing offsets of one structure
in the group.  Used for the type info table, and if "--gcremovetypeptr",
for all the vtables of the program as well.

The JIT does not support yet "--gcremovetypeptr" but that could be fixed.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4068314