Merge the branch gc-hash.  Cleans up the usage of hashes.  Now hash() is no longer RPython, and the official interface is defined in rlib/objectmodel: compute_hash() and compute_identity_hash().
Any GC object can now have its identity hash taken.  Simplifies code in
the PyPy interpreter, and allows a proper fix to be designed to replace
cast_ref_to_hashable() in the JIT.  "Small dicts" are gone now; the
other usages of hash() in RPython have been fixed.  A side effect is
that calling 'object.__hash__(x)' on top of pypy-c always works now.

On ootype, 'ooidentityhash' and 'oohash' are gone, replaced respectively
by 'identityhash' (which works both on lltype and ootype) and by a
method of the ootype.String class.

Adapt the GCs to add support for identityhash().  For moving GCs, this
is done by adding a 4-bytes field at the end of every object, *but only
when needed*.  This means that in the common case objects are 4 bytes
smaller, as long as we don't take their hash.  When we do, they grow by
4 bytes the next time the GC moves them.  This logic is not really
implemented yet for the mark&compact GC -- I disabled it right now.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4068537