Merge the change-celldict2 branch:
Strip the cell dict down to something radically simple. Now it is really just a
dict mapping names to cells. For the non-JIT case this makes global-lookups
non-optimized (but the caching seems to not have helped much anyway). For the
JIT-case the effect is nearly as good as the old code, and even better, if the
same key is looked up many times in the same module (which is quite common for
libraries).

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4070440