(pedronis, fijal) Merge the direct-assembler-call branch.
This branch implements a new operation CALL_ASSEMBLER, which happens
when we encounter a recursive portal call to a place that has assembler
already compiled. Speeds up certain cases, some by a significant %.
Slows down megamorphic calls, because with this branch, a call that was
compiled as CALL_ASSEMBLER is now a guard failure in case code object
is not the same.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4070829