Merge branch/fix-blackhole-virt.
This removes 'globaldata.blackhole_virtualizable',
which is not thread- or recursion-safe.

This is mostly a revert of r67161, but the test that r67161 tried to fix
just works nowadays, because we are doing "the right thing" about
virtualizables (it is correctly in the state TOKEN_NONE when
blackholing).  So I think that the test really just passes, and it would
not fail in more complicated cases either.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4071510