Merge branch/no-_immutable_.
Unlike hinted by the name of the branch, this does not kill the
_immutable_ hint; it just makes it saner (and _immutable_fields_ too).
Now, the definition in pure Python terms is precisely that an instance x
is fully immutable if and only if hasattr(x, '_immutable_').

During translation, it is now an error (reported by rpython/rclass.py)
if we find the _immutable_ hint on a class but, on some parent class, we
have non-trivial non-read-only attributes.

Similarly, it is an error to list a field name in _immutable_fields_ if
that field ends up on some parent class.

However, it is ok to give a field name that is not actually present in
this class.  If/when it shows up on some subclass, it will be flagged as
read-only there.  The idea is again that if we say
_immutable_fields_=['x'] on some class A, then any attribute 'x' on any
instance of (a subclass of) A will be read-only.

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4076857