(antocuni, arigo)
Hopefully fix lib-python/.../test_descr on 64-bit platforms.  The issue
is that we can get a cycle of tail recursions (more easily on x86-64
than on x86-32), and in such cycles, the stack overflow detection does
not work as a way to interrupt the cycle.  Thus we can get caught in
what seems like an infinite recursion (but is not in C).

--HG--
extra : convert_revision : svn%3A5fdb0b31-00ae-4507-bd76-94d651a7451c/pypy/trunk%4078133