Implement quasi-immutable fields which contain immutable lists, defined with _immutable_fields_ = ['foo?[*]'].
It actually means that 'foo' is a regular quasi-immutable field,
and that all list accesses we do directly by 'obj.foo[index]' are
supposed to read the **immutable** content of the list.  The
content of the list is not itself quasi-immutable.  Hence the
precise notation: 'lst?[*]', and not 'lst[*]?'.

--HG--
branch : out-of-line-guards-2